# Re-use jsonschema Validator instances across records and always use Draft4Validator